ui/requirements.txt
streamlit
python-dotenv
aiohttp